requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
python-dotenv
pydantic
langchain
langchain-google-genai
